+++ src/utils/collection.ts
@@ -1,0 +1,48 @@
+import { MediumGroup, SizeGroup } from '../types';
+
+export const deriveMediumGroup = (medium: string): MediumGroup => {
+  const lowercaseMedium = medium.toLowerCase();
+  if (lowercaseMedium.includes('acrylic')) return MediumGroup.Acrylic;
+  if (lowercaseMedium.includes('charcoal')) return MediumGroup.Charcoal;
+  if (lowercaseMedium.includes('collage')) return MediumGroup.Collage;
+  if (lowercaseMedium.includes('ink')) return MediumGroup.Ink;
+  if (lowercaseMedium.includes('mixed media'))
+    return MediumGroup['Mixed Media'];
+  if (lowercaseMedium.includes('monotype')) return MediumGroup.Monotype;
+  if (lowercaseMedium.includes('oil')) return MediumGroup.Oil;
+  if (lowercaseMedium.includes('pastel')) return MediumGroup.Pastel;
+  if (lowercaseMedium.includes('watercolor')) return MediumGroup.Watercolor;
+  return MediumGroup.Unknown;
+};
+
+const getArea = (dimensions: string): number => {
+  if (!dimensions.includes('"')) return 0;
+  const $dimensions = dimensions.replace('"', '').toLowerCase();
+
+  if ($dimensions.includes('x')) {
+    const [h, w] = $dimensions.split('x');
+    const area = Number(h) * Number(w);
+    if (Number.isNaN(area)) return 0;
+    return area;
+  }
+
+  if ($dimensions.includes('diameter')) {
+    const [h] = $dimensions.split('diameter');
+    const area = Number(h) * Number(h);
+    if (Number.isNaN(area)) return 0;
+    return area;
+  }
+
+  return 0;
+};
+
+export const deriveSizeGroup = (dimensions: string): SizeGroup => {
+  const area = getArea(dimensions);
+
+  if (area === 0) return SizeGroup.Unknown;
+  if (area < 75) return SizeGroup['Very Small'];
+  if (area < 200) return SizeGroup.Small;
+  if (area < 500) return SizeGroup.Medium;
+  if (area < 1000) return SizeGroup.Large;
+  return SizeGroup['Very Large'];
+};
