+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './arrays';
+export * from './collection';
